backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.13"
dependencies = [
    "alembic>=1.16.5",
    "fastapi>=0.117.1",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.12",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.3.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.36.0",
]
